refactor: override to string methods
This overrides the to string methods to the static and dynamic array classes so that it can print out the elements in each array.
This also repackages both arrays into its own separate packages.

author(s): [Emmanuel]
closes: #10
